testing pwa
to support mobile gamers, we wanted to create a mini app on telegram, but it proved a little bit challenging with standard wallet connections since miniapps works inside of an iframe. we created a backend wallet solution but its a setup that require robust security. We will be focusing on pwa instead as it just works out of the box without complex setup.